Refactor HasSchemaPrivilegeFunction and extract superclass
As a preparation step for `HasDatabasePrivilegeFunction` extract common
parts to a superclass `HasPrivilegeFunction`.
